refactor(codec): own zoned encoding policy
