+++ crates/copybook-codec/Cargo.toml
@@ -24,7 +24,6 @@ targets = ["x86_64-unknown-linux-gnu"]
 [dependencies]
 copybook-core = { workspace = true }
 copybook-charset.workspace = true
-copybook-zoned-format.workspace = true
 copybook-overpunch.workspace = true
 copybook-fixed.workspace = true
 copybook-rdw.workspace = true
