+++ crates/copybook-zoned-format/Cargo.toml
@@ -21,7 +21,7 @@ all-features = false
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
-serde.workspace = true
+copybook-codec.workspace = true
 
 [dev-dependencies]
 proptest.workspace = true
